real_survillance/backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
opencv-python
requests
pyjwt
facenet-pytorch
torch
torchvision
torchaudio
python-dotenv>=1.0.0
twilio>=8.0.0
